Move FulNodeBuilder available services test to FullNodeBuilderTest
